[WIP]Copied android UI to common
 -> Desktop use same UI
