+++ common/src/commonMain/kotlin/com/github/kentvu/t9vietnamese/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.github.kentvu.t9vietnamese.android.ui.theme
+package com.github.kentvu.t9vietnamese.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
