+++ common/src/commonMain/kotlin/com/github/kentvu/t9vietnamese/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.github.kentvu.t9vietnamese.android.ui.theme
+package com.github.kentvu.t9vietnamese.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
